+++ CHANGELOG.md
@@ -8,6 +8,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 The intended audience of this file is for py42 consumers -- as such, changes that don't affect
 how a consumer would use the library (e.g. adding unit tests, updating documentation, etc) are not captured here.
 
+## 1.5.1 - 2020-06-17
+
+### Fixed
+
+- An unintended regression that caused `import py42.sdk.settings.debug` to fail.
+
+### Changed
+
+- For security reasons, debug output no longer logs the headers on http requests.
+
 ## 1.5.0 - 2020-06-16
 
 ### Added